ui: replace stream selection tables
Use `display: grid;` instead of actual `<table>` elements.

These grids are composed of labels which have a single `<input>`
sibling and one (or more) `<span>` siblings. When the radio button
is selected we highlight the whole row; furthermore the entire row
is considered a label for the purposes of input handling.